Split classes into own files from review
