+++ oop/Dog.php
@@ -1,0 +1,9 @@
+<?php
+
+class Dog extends Pet {
+
+    public function makeNoise()
+    {
+        echo "BARK!";
+    }
+}
