+++ oop/Goat.php
@@ -1,0 +1,9 @@
+<?php
+
+class Goat extends Pet {
+
+    public function makeNoise()
+    {
+        // TODO: Implement makeNoise() method.
+    }
+}
