+++ oop/Pet.php
@@ -1,5 +1,9 @@
 <?php
 
+require_once 'Cat.php';
+require_once 'Dog.php';
+require_once 'Goat.php';
+
 abstract class Pet {
 
     private $owner;
@@ -20,38 +24,6 @@ abstract class Pet {
     }
 
     public abstract function makeNoise();
-}
-
-class Cat extends Pet {
-
-    public function setOwner($owner) {
-        if ($owner !== 'jeff') {
-            throw new Exception('You cant own a cat unless you are jeff');
-        }
-
-        parent::setOwner($owner);
-    }
-
-    public function makeNoise()
-    {
-        echo 'MEOWWWW!';
-    }
-}
-
-class Dog extends Pet {
-
-    public function makeNoise()
-    {
-        echo "BARK!";
-    }
-}
-
-class Goat extends Pet {
-
-    public function makeNoise()
-    {
-        // TODO: Implement makeNoise() method.
-    }
 }
 
 $cat = new Cat("dave");
